implemented wireless communication
using the WiFi library from Arduino, implemented a simple client server system with two ESP32 boards. when a push button connected to the client board is pressed, it prints data to the server board that it's connected to that then prints the state of the button on an lcd screen. I used an lcd screen to output data since i'm unable to monitor two ports at the same time